add login_view, move main to teddy_view, and try to do main | TODO: fix main.py
